DBFilehandler added

+++ project_2/src/utilities/Doctor.java
@@ -1,7 +1,7 @@
 package utilities;
 
 public class Doctor extends User {
-	public Doctor(String username, String password, String division, int certNbr) {
+	public Doctor(String username, String password, String division, long certNbr) {
 		super(username, password, division, certNbr);
 		this.permLevel = PermissionLevel.Doctor;
 	}
